add wp-ord assets
