--- assets/css/styleCSS/css/paymentPlan.css
@@ -1,5 +1,0 @@
-@tailwind base;
-
-@tailwind components;
-
-@tailwind utilities;
