--- assets/css/styleCSS/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "rave-payment-form",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "autoprefixer": "^9.8.8",
-    "postcss": "^7.0.36",
-    "tailwindcss": "^2.2.0"
-  }
-}
