--- assets/css/styleCSS/tailwind.config.js
@@ -1,16 +1,0 @@
-module.exports = {
-  future: {
-    // removeDeprecatedGapUtilities: true,
-    // purgeLayersByDefault: true,
-  },
-  purge: [],
-  theme: {
-    extend: {
-
-    },
-  },
-  variants: {
-    tableLayout: ['responsive', 'hover', 'focus'],
-  },
-  plugins: [],
-}
