+++ includes/class-flw-payment-list.php
@@ -1,23 +1,25 @@
 <?php
-  /**
-   * Rave Payment List
-   */
-  if ( ! defined( 'ABSPATH' ) ) { exit; }
+/**
+ * Flutterwave Payment List
+ *
+ * @package Flutterwave Payment
+ */
 
-  // require_once( ABSPATH . 'wp-admin/includes/screen.php');
+defined( 'ABSPATH' ) || exit;
+
   require_once( ABSPATH . 'wp-admin/includes/template.php');
   require_once( ABSPATH . 'wp-admin/includes/class-wp-screen.php');
   if( ! class_exists( 'FLW_WP_List_Table' ) ) {
     require_once( FLW_DIR_PATH . 'includes/wp-classes/class-wp-list-table.php' );
   }
 
-  if ( ! class_exists( 'FLW_Rave_Payment_List' ) ) {
+  if ( ! class_exists( 'FLW_Payment_List' ) ) {
 
     /**
      * Payment List Class to add list payments made
      * via the payment buttons
      */
-    class FLW_Rave_Payment_List extends FLW_WP_List_Table {
+    class FLW_Payment_List extends FLW_WP_List_Table {
 
       /**
        * Class Instance
